+++ src/test/java/gov/usgs/wma/waterdata/timeseries/TimeSeriesController1095IT.java
@@ -1,7 +1,6 @@
 package gov.usgs.wma.waterdata.timeseries;
 
 import com.github.springtestdbunit.annotation.DatabaseSetup;
-import gov.usgs.wma.waterdata.springinit.BaseIT;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
@@ -16,8 +15,11 @@ import java.io.IOException;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
 import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONObjectAs;
+
+import gov.usgs.wma.waterdata.springinit.BaseIT;
 
 /**
  * The '1095' tests specifically tests that timeseries data older than 1095 days is not displayed if not approved.
@@ -26,7 +28,7 @@ import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONObjectAs;
  * separate test classes with '1095' in the name.
  */
 @SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT)
-@DatabaseSetup("classpath:/testData/monitoringLocation/")
+@DatabaseSetup("classpath:/testData/timeSeries/")
 @DatabaseSetup("classpath:/testData/groundwaterDailyValue1095/")
 public class TimeSeriesController1095IT extends BaseIT {
 
@@ -35,8 +37,27 @@ public class TimeSeriesController1095IT extends BaseIT {
 
 	@Test
 	public void foundTest() {
-		ResponseEntity<String> rtn = restTemplate.getForEntity("/monitoring-location/USGS-07227448/time-series/e6a4cc2de5bf437e83efe0107cf026ac", String.class);
+		String url = "/collections/monitoring-locations/items/USGS-07227448/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
 		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		assertNotNull(rtn.getBody());
+
+		try {
+			assertThat(new JSONObject(rtn.getBody()),
+					sameJSONObjectAs(new JSONObject(getCompareFile("e6a4cc2de5bf437e83efe0107cf026ac_1095.json"))));
+		} catch (JSONException e) {
+			fail("Unexpected JSONException during test", e);
+		} catch (IOException e) {
+			fail("Unexpected IOException during test", e);
+		}
+	}
+
+	@Test
+	public void foundInSatCollectionTest() {
+		String url = "/collections/SAT/items/USGS-07227448/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		assertNotNull(rtn.getBody());
 
 		try {
 			assertThat(new JSONObject(rtn.getBody()),