+++ src/test/java/gov/usgs/wma/waterdata/timeseries/TimeSeriesControllerIT.java
@@ -2,7 +2,8 @@ package gov.usgs.wma.waterdata.timeseries;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
 import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONObjectAs;
 
@@ -22,8 +23,8 @@ import com.github.springtestdbunit.annotation.DatabaseSetup;
 
 import gov.usgs.wma.waterdata.springinit.BaseIT;
 
-@SpringBootTest(webEnvironment=WebEnvironment.RANDOM_PORT)
-@DatabaseSetup("classpath:/testData/monitoringLocation/")
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@DatabaseSetup("classpath:/testData/timeSeries/")
 @DatabaseSetup("classpath:/testData/groundwaterDailyValue/")
 public class TimeSeriesControllerIT extends BaseIT {
 
@@ -32,8 +33,27 @@ public class TimeSeriesControllerIT extends BaseIT {
 
 	@Test
 	public void foundTest() {
-		ResponseEntity<String> rtn = restTemplate.getForEntity("/monitoring-location/USGS-07227448/time-series/e6a4cc2de5bf437e83efe0107cf026ac", String.class);
+		String url = "/collections/monitoring-locations/items/USGS-07227448/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
 		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		assertNotNull(rtn.getBody());
+
+		try {
+			assertThat(new JSONObject(rtn.getBody()),
+					sameJSONObjectAs(new JSONObject(getCompareFile("e6a4cc2de5bf437e83efe0107cf026ac.json"))));
+		} catch (JSONException e) {
+			fail("Unexpected JSONException during test", e);
+		} catch (IOException e) {
+			fail("Unexpected IOException during test", e);
+		}
+	}
+
+	@Test
+	public void foundInSatCollectionTest() {
+		String url = "/collections/SAT/items/USGS-07227448/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.OK));
+		assertNotNull(rtn.getBody());
 
 		try {
 			assertThat(new JSONObject(rtn.getBody()),
@@ -47,15 +67,33 @@ public class TimeSeriesControllerIT extends BaseIT {
 
 	@Test
 	public void notFoundTest() {
-		ResponseEntity<String> rtn = restTemplate.getForEntity("/monitoring-location/USGS-12345678/time-series/216d009de8914147a0f9e5237da77854", String.class);
+		String url = "/collections/monitoring-locations/items/USGS-12345678/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
 		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
-		assertNull(rtn.getBody());
+		assertEquals(ogc404Payload, rtn.getBody());
+	}
+
+	@Test
+	public void notFoundInSatCollectionTest() {
+		String url = "/collections/monitoring-locations/items/USGS-12345678/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+		assertEquals(ogc404Payload, rtn.getBody());
 	}
 
 	@Test
 	public void noGeomTest() {
-		ResponseEntity<String> rtn = restTemplate.getForEntity("/monitoring-location/USGS-04028090/time-series/41a5ff887b744b84a271b65e48d78074", String.class);
+		String url = "/collections/monitoring-locations/items/USGS-04028090/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
 		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
-		assertNull(rtn.getBody());
+		assertEquals(ogc404Payload, rtn.getBody());
+	}
+
+	@Test
+	public void noGeomSatCollectionTest() {
+		String url = "/collections/SAT/items/USGS-04028090/observations/statistical-time-series/e6a4cc2de5bf437e83efe0107cf026ac";
+		ResponseEntity<String> rtn = restTemplate.getForEntity(url, String.class);
+		assertThat(rtn.getStatusCode(), equalTo(HttpStatus.NOT_FOUND));
+		assertEquals(ogc404Payload, rtn.getBody());
 	}
 }